+++ components/catalogs/back/tests/Integration/Infrastructure/Validation/ProductSelection/AttributeCriterion/AttributeTextareaCriterionTest.php
@@ -219,7 +219,7 @@ class AttributeTextareaCriterionTest extends AbstractAttributeCriterionTest
     public function testItReturnsViolationsWhenInvalid(
         array $attribute,
         array $criterion,
-        string $expectedMessage
+        string $expectedMessage,
     ): void {
         $this->createAttribute($attribute);