+++ components/identifier-generator/front/src/feature/tabs/structure/edit/FamilyPropertyEdit.tsx
@@ -5,6 +5,7 @@ import {Field, NumberInput, SelectInput} from 'akeneo-design-system';
 import {useTranslate} from '@akeneo-pim-community/shared';
 import {NomenclatureEdit, OperatorSelector} from '../../../components';
 import {Styled} from '../../../components/Styled';
+import {useIdentifierGeneratorAclContext} from '../../../context';
 
 const options = [
   {value: AbbreviationType.TRUNCATE, label: 'pim_identifier_generator.structure.settings.code_format.type.truncate'},
@@ -17,6 +18,7 @@ const options = [
 
 const FamilyPropertyEdit: PropertyEditFieldsProps<FamilyProperty> = ({selectedProperty, onChange}) => {
   const translate = useTranslate();
+  const identifierGeneratorAclContext = useIdentifierGeneratorAclContext();
 
   const onChangeProcessType = useCallback(
     (type: string) => {
@@ -84,6 +86,7 @@ const FamilyPropertyEdit: PropertyEditFieldsProps<FamilyProperty> = ({selectedPr
           openLabel={translate('pim_common.open')}
           onChange={onChangeProcessType}
           clearable={false}
+          readOnly={!identifierGeneratorAclContext.isManageIdentifierGeneratorAclGranted}
         >
           {options.map(({value, label}) => (
             <SelectInput.Option value={value} title={translate(label)} key={value}>
@@ -108,7 +111,13 @@ const FamilyPropertyEdit: PropertyEditFieldsProps<FamilyProperty> = ({selectedPr
             label={translate('pim_identifier_generator.structure.settings.family.chars_number')}
             requiredLabel={translate('pim_common.required_label')}
           >
-            <NumberInput value={`${selectedProperty.process.value}`} onChange={onChangeCharsNumber} max={5} min={1} />
+            <NumberInput
+              value={`${selectedProperty.process.value}`}
+              onChange={onChangeCharsNumber}
+              max={5}
+              min={1}
+              readOnly={!identifierGeneratorAclContext.isManageIdentifierGeneratorAclGranted}
+            />
           </Field>
         </>
       )}