+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Controller/Internal/GetConnectedAppAction.php
@@ -51,11 +51,7 @@ final class GetConnectedAppAction
 
     private function denyAccessUnlessGrantedToManage(ConnectedApp $connectedApp): void
     {
-        if (!$connectedApp->isTestApp() && !$this->security->isGranted('akeneo_connectivity_connection_manage_apps')) {
-            throw new AccessDeniedHttpException();
-        }
-
-        if ($connectedApp->isTestApp() && !$this->security->isGranted('akeneo_connectivity_connection_manage_test_apps')) {
+        if (!$connectedApp->isCustomApp() && !$this->security->isGranted('akeneo_connectivity_connection_manage_apps')) {
             throw new AccessDeniedHttpException();
         }
     }